backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
groq==0.4.2
httpx==0.27.0
python-dotenv==1.0.1
alembic==1.13.1
PyJWT[crypto]>=2.8.0
